pyproject.toml
[project]
name = "ui-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dora-rs>=0.3.8",
    "gradio>=5.3.0",
    "robosuite>=1.5.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
